+++ src/constants/app.js
@@ -4,13 +4,8 @@ export const ROUTE_APP_PRODUCTION_DOMAIN_PROTOCOL = "https://"
 export const APP_NAME = "Analog.Cafe"
 export const APP_DESCRIPTION = "A film photography publication"
 
-// api location
-export const WEBSOCKET_UPLOAD_PROGRESS =
-  process.env.REACT_APP_WEBSOCKET_UPLOAD_PROGRESS
-export const WEBSOCKET_AUTH_TOKEN = process.env.REACT_APP_WEBSOCKET_AUTH_TOKEN
-export const ROUTE_API_DOMAIN = process.env.REACT_APP_API_DOMAINSS
-
-// NOTE: this doesn't work yet
+// key routes
+export const ROUTE_API_DOMAIN = process.env.REACT_APP_API_DOMAIN
 export const ROUTE_APP_CURRENT_DOMAIN = window.location.hostname
 
 // Google Analytics
