+++ src/constants/user.js
@@ -1,3 +1,6 @@
 // api endpoint
 import { ROUTE_API_DOMAIN } from "./app"
 export const ROUTE_USER_API = ROUTE_API_DOMAIN + "/api/auth/user"
+
+// WebSocket
+export const WEBSOCKET_AUTH_TOKEN = process.env.REACT_APP_WEBSOCKET_AUTH_TOKEN
